[IMP]pms_fastapi: Added id-number-categories endpoint.
