+++ pms_fastapi/schemas/__init__.py
@@ -7,3 +7,4 @@ from . import country_state
 from . import payment_term
 from . import pricelist
 from . import contact_id_number
+from . import contact_tag
